cohorts/2024/MLZoomcampCapstone2/requirements.txt
flask>=3.1.0
numpy>=2.2.2
pandas>=2.2.3
pillow==10.0.0
scikit-learn>=1.6.1
streamlit>=1.41.1
tensorflow>=2.14.0
torch>=2.5.1
torchvision>=0.20.1
tqdm>=4.67.1
